requirements.txt
asgiref==3.5.2
Django==4.1.4
django-environ==0.9.0
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
